Refactor: move setup to manager hook
