+++ src/Manager/Manager.tsx
@@ -10,14 +10,24 @@ import ButtonToolbar from 'react-bootstrap/ButtonToolbar';
 import { useDispatch, useSelector } from 'react-redux';
 import {
     AppThunk,
+    isDevelopment,
+    isLoggingVerbose,
     logger,
     usageData,
 } from '@nordicsemiconductor/pc-nrfconnect-shared';
 
 import FirstInstallInstructions from '../FirstInstall/FirstInstallInstructions';
 import InstallDirDialog from '../InstallDir/InstallDirDialog';
+import {
+    currentInstallDir,
+    setInstallDir,
+} from '../InstallDir/installDirSlice';
 import InstallPackageDialog from '../InstallPackageDialog/InstallPackageDialog';
 import NrfCard from '../NrfCard/NrfCard';
+import {
+    persistedInstallDir,
+    setPersistedInstallDir,
+} from '../persistentStore';
 import ReduxConfirmDialog from '../ReduxConfirmDialog/ReduxConfirmDialog';
 import { isOlderEnvironmentsHidden } from '../Settings/settingsSlice';
 import { RootState } from '../state';
@@ -33,6 +43,7 @@ import { VsCodeStatus } from '../VsCodeDialog/vscodeSlice';
 import detectMultipleInstallDirs from './detectMultipleInstallDirs';
 import Environment from './Environment/Environment';
 import RemoveEnvironmentDialog from './Environment/RemoveEnvironmentDialog';
+import initEnvironments from './initEnvironments';
 import {
     environmentsByVersion,
     isShowingFirstSteps,
@@ -40,6 +51,7 @@ import {
 } from './managerSlice';
 import NrfUtilEnvDialog from './nrfutil/NrfUtilDialog';
 import PlatformInstructions from './PlatformInstructions';
+import toolchainManager from './ToolchainManager/toolchainManager';
 import { filterEnvironments } from './versionFilter';
 
 const Environments = () => {
@@ -70,12 +82,38 @@ const Environments = () => {
     );
 };
 
-export default () => {
+const useManagerHooks = () => {
     const dispatch = useDispatch();
+    const verboseLogging = useSelector(isLoggingVerbose);
+    const installDir = useSelector(currentInstallDir);
+
     useEffect(() => {
         dispatch(initApp());
+    }, [dispatch, installDir]);
+
+    useEffect(() => {
+        const fallback = isDevelopment ? 'error' : 'off';
+        toolchainManager.setLogLevel(verboseLogging ? 'trace' : fallback);
+    }, [verboseLogging]);
+
+    useEffect(() => {
+        if (!persistedInstallDir()) {
+            toolchainManager.config().then(config => {
+                setPersistedInstallDir(config.install_dir);
+                dispatch(setInstallDir(config.install_dir));
+            });
+        }
     }, [dispatch]);
 
+    useEffect(() => {
+        dispatch(initEnvironments());
+    }, [dispatch, installDir]);
+};
+
+export default () => {
+    const dispatch = useDispatch();
+
+    useManagerHooks();
     const showingFirstSteps = useSelector(isShowingFirstSteps);
 
     if (showingFirstSteps) {
